CMMS_Project/requirements.txt
# UI Framework
flet==0.23.2

# Database
sqlalchemy==2.0.23
alembic==1.13.1
pymysql==1.1.0

# Data Processing
pandas==2.1.3
openpyxl>=3.1.5
matplotlib==3.8.2

# File & Image Handling
qrcode==7.4.2
Pillow==10.1.0
python-magic==0.4.27

# PDF Generation
reportlab==4.0.9
python-docx==1.1.0
docx2pdf==0.1.8

# Security & Cryptography
# Note: argon2-cffi removed - using bcrypt via passlib (GDPR compliance requirement)
cryptography==41.0.7

# Localization
babel==2.14.0

# REST API
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.3

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
requests==2.31.0
schedule==1.2.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Development
black==23.12.0
flake8==6.1.0
isort==5.13.2

# Packaging
pyinstaller==6.3.0
